Reorganize file structure for better clarity
This commit restructures the project by moving several files and updating their corresponding imports.

- Moved `Controller.java` to a new `software` package.
- Moved `UnifiedLogging.java` to the `configuration` package.
- Relocated all data collection and testing scripts from the `data` directory to a new `testing/data` directory.
- Updated the `.gitignore` file to reflect the new path for the Python virtual environment.
- Adjusted import statements across `MainAuto.java`, `MainOp.java`, and `Settings.java` to match the new file locations.
