+++ .gitignore
@@ -91,4 +91,4 @@ lint/tmp/
 
 
 # data collection venv
-/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/data/.venv/
+/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/data/.venv/
