+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/configuration/Settings.java
@@ -4,7 +4,7 @@ import com.bylazar.configurables.annotations.Configurable;
 import com.pedropathing.geometry.BezierCurve;
 import com.pedropathing.geometry.Pose;
 
-import org.firstinspires.ftc.teamcode.Controller;
+import org.firstinspires.ftc.teamcode.software.Controller;
 
 import java.util.EnumMap;
