+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/configuration/UnifiedLogging.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.configuration;
 
 import com.bylazar.telemetry.TelemetryManager;
 import com.pedropathing.follower.Follower;
