+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/software/Controller.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.software;
 
 import com.pedropathing.follower.Follower;
 import com.pedropathing.math.Vector;
